# isme-nest-serrve.swagger.json
{
	"info": {
		"title": "isme-nest-serrve",
		"description": "",
		"version": "1.0.0"
	},
...
	"paths": {
...
		"/role/permissions/tree": {
			"get": {
				"summary": "角色权限树-by token",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"name": {
												"type": "string"
											},
											"code": {
												"type": "string"
											},
											"type": {
												"type": "string"
											},
											"parentId": {
												"type": "null"
											},
											"path": {
												"type": "string",
												"x-nullable": true
											},
											"redirect": {
												"type": "null"
											},
											"icon": {
												"type": "string"
											},
											"component": {
												"type": "string",
												"x-nullable": true
											},
											"layout": {
												"type": "string",
												"x-nullable": true
											},
											"keepAlive": {
												"type": "null"
											},
											"method": {
												"type": "null"
											},
											"description": {
												"type": "null"
											},
											"show": {
												"type": "boolean"
											},
											"enable": {
												"type": "boolean"
											},
											"order": {
												"type": "integer"
											},
											"children": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"id": {
															"type": "integer"
														},
														"name": {
															"type": "string"
														},
														"code": {
															"type": "string"
														},
														"type": {
															"type": "string"
														},
														"parentId": {
															"type": "integer"
														},
														"path": {
															"type": "string"
														},
														"redirect": {
															"type": "null"
														},
														"icon": {
															"type": "string"
														},
														"component": {
															"type": "string"
														},
														"layout": {
															"type": "string",
															"x-nullable": true
														},
														"keepAlive": {
															"type": "boolean",
															"x-nullable": true
														},
														"method": {
															"type": "null"
														},
														"description": {
															"type": "null"
														},
														"show": {
															"type": "boolean"
														},
														"enable": {
															"type": "boolean"
														},
														"order": {
															"type": "integer"
														},
														"children": {
															"type": "array",
															"items": {
																"type": "object",
																"properties": {
																	"id": {
																		"type": "integer"
																	},
																	"name": {
																		"type": "string"
																	},
																	"code": {
																		"type": "string"
																	},
																	"type": {
																		"type": "string"
																	},
																	"parentId": {
																		"type": "integer"
																	},
																	"path": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"redirect": {
																		"type": "null"
																	},
																	"icon": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"component": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"layout": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"keepAlive": {
																		"type": "null"
																	},
																	"method": {
																		"type": "null"
																	},
																	"description": {
																		"type": "null"
																	},
																	"show": {
																		"type": "boolean"
																	},
																	"enable": {
																		"type": "boolean"
																	},
																	"order": {
																		"type": "integer"
																	}
																},
																"required": ["id", "name", "code", "type", "parentId",
																	"path", "redirect", "icon", "component",
																	"layout", "keepAlive", "method", "description",
																	"show", "enable", "order"
																]
															}
														}
													},
													"required": ["id", "name", "code", "type", "parentId", "path",
														"redirect", "icon", "component", "layout", "keepAlive",
														"method", "description", "show", "enable", "order",
														"children"
													]
												}
											}
										},
										"required": ["id", "name", "code", "type", "parentId", "path", "redirect",
											"icon", "component", "layout", "keepAlive", "method", "description",
											"show", "enable", "order", "children"
										]
									}
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/permission/menu/tree": {
			"get": {
				"summary": "权限树-菜单",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"name": {
												"type": "string"
											},
											"code": {
												"type": "string"
											},
											"type": {
												"type": "string"
											},
											"parentId": {
												"type": "null"
											},
											"path": {
												"type": "string",
												"x-nullable": true
											},
											"redirect": {
												"type": "null"
											},
											"icon": {
												"type": "string"
											},
											"component": {
												"type": "string",
												"x-nullable": true
											},
											"layout": {
												"type": "string",
												"x-nullable": true
											},
											"keepAlive": {
												"type": "null"
											},
											"method": {
												"type": "null"
											},
											"description": {
												"type": "null"
											},
											"show": {
												"type": "boolean"
											},
											"enable": {
												"type": "boolean"
											},
											"order": {
												"type": "integer"
											},
											"children": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"id": {
															"type": "integer"
														},
														"name": {
															"type": "string"
														},
														"code": {
															"type": "string"
														},
														"type": {
															"type": "string"
														},
														"parentId": {
															"type": "integer"
														},
														"path": {
															"type": "string"
														},
														"redirect": {
															"type": "null"
														},
														"icon": {
															"type": "string"
														},
														"component": {
															"type": "string"
														},
														"layout": {
															"type": "string",
															"x-nullable": true
														},
														"keepAlive": {
															"type": "boolean",
															"x-nullable": true
														},
														"method": {
															"type": "null"
														},
														"description": {
															"type": "null"
														},
														"show": {
															"type": "boolean"
														},
														"enable": {
															"type": "boolean"
														},
														"order": {
															"type": "integer"
														},
														"children": {
															"type": "array",
															"items": {
																"type": "object",
																"properties": {
																	"id": {
																		"type": "integer"
																	},
																	"name": {
																		"type": "string"
																	},
																	"code": {
																		"type": "string"
																	},
																	"type": {
																		"type": "string"
																	},
																	"parentId": {
																		"type": "integer"
																	},
																	"path": {
																		"type": "string"
																	},
																	"redirect": {
																		"type": "null"
																	},
																	"icon": {
																		"type": "string"
																	},
																	"component": {
																		"type": "string"
																	},
																	"layout": {
																		"type": "string"
																	},
																	"keepAlive": {
																		"type": "null"
																	},
																	"method": {
																		"type": "null"
																	},
																	"description": {
																		"type": "null"
																	},
																	"show": {
																		"type": "boolean"
																	},
																	"enable": {
																		"type": "boolean"
																	},
																	"order": {
																		"type": "integer"
																	}
																}
															}
														}
													},
													"required": ["id", "name", "code", "type", "parentId", "path",
														"redirect", "icon", "component", "layout", "keepAlive",
														"method", "description", "show", "enable", "order"
													]
												}
											}
										},
										"required": ["id", "name", "code", "type", "parentId", "path", "redirect",
											"icon", "component", "layout", "keepAlive", "method", "description",
											"show", "enable", "order", "children"
										]
									}
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/permission/tree": {
			"get": {
				"summary": "权限树-all",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"name": {
												"type": "string"
											},
											"code": {
												"type": "string"
											},
											"type": {
												"type": "string"
											},
											"parentId": {
												"type": "null"
											},
											"path": {
												"type": "string",
												"x-nullable": true
											},
											"redirect": {
												"type": "null"
											},
											"icon": {
												"type": "string"
											},
											"component": {
												"type": "string",
												"x-nullable": true
											},
											"layout": {
												"type": "string",
												"x-nullable": true
											},
											"keepAlive": {
												"type": "null"
											},
											"method": {
												"type": "null"
											},
											"description": {
												"type": "null"
											},
											"show": {
												"type": "boolean"
											},
											"enable": {
												"type": "boolean"
											},
											"order": {
												"type": "integer"
											},
											"children": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"id": {
															"type": "integer"
														},
														"name": {
															"type": "string"
														},
														"code": {
															"type": "string"
														},
														"type": {
															"type": "string"
														},
														"parentId": {
															"type": "integer"
														},
														"path": {
															"type": "string"
														},
														"redirect": {
															"type": "null"
														},
														"icon": {
															"type": "string"
														},
														"component": {
															"type": "string"
														},
														"layout": {
															"type": "string",
															"x-nullable": true
														},
														"keepAlive": {
															"type": "boolean",
															"x-nullable": true
														},
														"method": {
															"type": "null"
														},
														"description": {
															"type": "null"
														},
														"show": {
															"type": "boolean"
														},
														"enable": {
															"type": "boolean"
														},
														"order": {
															"type": "integer"
														},
														"children": {
															"type": "array",
															"items": {
																"type": "object",
																"properties": {
																	"id": {
																		"type": "integer"
																	},
																	"name": {
																		"type": "string"
																	},
																	"code": {
																		"type": "string"
																	},
																	"type": {
																		"type": "string"
																	},
																	"parentId": {
																		"type": "integer"
																	},
																	"path": {
																		"type": "string"
																	},
																	"redirect": {
																		"type": "null"
																	},
																	"icon": {
																		"type": "string"
																	},
																	"component": {
																		"type": "string"
																	},
																	"layout": {
																		"type": "string"
																	},
																	"keepAlive": {
																		"type": "null"
																	},
																	"method": {
																		"type": "null"
																	},
																	"description": {
																		"type": "null"
																	},
																	"show": {
																		"type": "boolean"
																	},
																	"enable": {
																		"type": "boolean"
																	},
																	"order": {
																		"type": "integer"
																	}
																}
															}
														}
													},
													"required": ["id", "name", "code", "type", "parentId", "path",
														"redirect", "icon", "component", "layout", "keepAlive",
														"method", "description", "show", "enable", "order"
													]
												}
											}
										},
										"required": ["id", "name", "code", "type", "parentId", "path", "redirect",
											"icon", "component", "layout", "keepAlive", "method", "description",
											"show", "enable", "order", "children"
										]
									}
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/permission": {
			"post": {
				"summary": "新增权限",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/permission/{id}": {
			"delete": {
				"summary": "删除权限",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			},
			"patch": {
				"summary": "修改权限",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/permission/button/{parentId}": {
			"get": {
				"summary": "按钮权限-by parentId",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [{
					"name": "parentId",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"name": {
												"type": "string"
											},
											"code": {
												"type": "string"
											},
											"type": {
												"type": "string"
											},
											"parentId": {
												"type": "integer"
											},
											"path": {
												"type": "null"
											},
											"redirect": {
												"type": "null"
											},
											"icon": {
												"type": "null"
											},
											"component": {
												"type": "null"
											},
											"layout": {
												"type": "null"
											},
											"keepAlive": {
												"type": "null"
											},
											"method": {
												"type": "null"
											},
											"description": {
												"type": "null"
											},
											"show": {
												"type": "boolean"
											},
											"enable": {
												"type": "boolean"
											},
											"order": {
												"type": "integer"
											}
										}
									}
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
...
	},
	"swagger": "2.0",
	"basePath": null,
...
	"securityDefinitions": {
		"bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
...
}